feat: add generic tracking endpoint
