+++ backend/src/tracking/tracking.ts
@@ -1,44 +1,32 @@
-import cors from "cors";
 import { Request, Response, Router } from "express";
 
+import { BadRequestError } from "~backend/src/errors/errorTypes";
+import { getUserIdFromRequest } from "~backend/src/utils";
 import { trackBackendUserEvent } from "~shared/backendAnalytics";
 import { log } from "~shared/logger";
+import { AnalyticsEventName } from "~shared/types/analytics";
 
 import { HttpStatus } from "../http";
 
 export const router = Router();
 
-interface TrackingPayload {
-  userId: string;
-}
-
-/**
- * We're calling those endpoints from landing which is a different url, this we need to enable CORS for this endpoints
- * server side
- */
-router.use(
-  cors({
-    origin: [`https://acape.la`, `https://acapela.com`],
-  })
-);
-
 /**
  * This endpoint handles user signup calls from the landing page
  */
-router.post("/v1/setup", async (req: Request, res: Response) => {
-  const { userId } = req.body as TrackingPayload;
+router.post("/v1/track", async (req: Request, res: Response) => {
+  const userId = getUserIdFromRequest(req);
 
-  if (!userId) {
-    log.info("Tracking endpoint called with missing parameters");
-    return res.status(HttpStatus.BAD_REQUEST).end();
+  const eventName = req.body.eventName as AnalyticsEventName;
+  if (!eventName) {
+    throw new BadRequestError("event name missing");
   }
 
   try {
-    trackBackendUserEvent(userId, "Opened App");
+    trackBackendUserEvent(userId, eventName, req.body.payload);
     res.status(HttpStatus.OK).end();
   } catch (e) {
     console.error(e);
-    log.error(`Tracking a frontend event ${event} failed`);
+    log.error(`Tracking a frontend event ${eventName} failed`);
     return res.status(HttpStatus.CONFLICT).end();
   }
 });
