+++ backend/src/utils.ts
@@ -1,7 +1,10 @@
 import { Request, Response } from "express";
+import { get } from "lodash";
+
+import { verifyJWT } from "~shared/jwt";
 
 import { extractAndAssertBearerToken } from "./authentication";
-import { AuthenticationError } from "./errors/errorTypes";
+import { AuthenticationError, BadRequestError } from "./errors/errorTypes";
 
 export function middlewareRequireBearerToken(secretValue: string, errorMessage: string) {
   return function (req: Request, _: Response, next: () => unknown): void {
@@ -36,4 +39,25 @@ export function assertDate(dateString: unknown, message = "Incorrect date string
   if (!isValidOptionalDateArgument(dateString as string)) {
     throw new Error(message);
   }
+}
+
+export function getUserIdFromRequest(req: Request): string {
+  const jwtToken = req.cookies["next-auth.session-token"];
+
+  if (!jwtToken) {
+    throw new BadRequestError("session token missing");
+  }
+
+  let session;
+  try {
+    session = verifyJWT(jwtToken);
+  } catch (e) {
+    throw new AuthenticationError();
+  }
+
+  const userId = get(session, "id");
+  if (!userId) {
+    throw new BadRequestError("user id missing");
+  }
+  return userId;
 }
